chore: refactor code and improve error handling
- Removed unnecessary macro usage in `lib.rs`.
- Simplified error conversion in `error.rs`.
- Added unused macro allowance in `crud_handlers.rs`.
- Created a script to update git submodules.
- Updated submodule reference for `openrouter_api`.
- Cleaned up code in `openrouter.rs` and `lib.rs` for better readability.
- Enhanced database connection error handling in `lib.rs`.
- Added default implementations for repository structs.
- Refactored download service for improved clarity and performance.
- Updated various mappers in fetcher modules for consistency.
- Improved tag handling in `file.rs` and added default implementations for tagger providers.
- Cleaned up test files for better maintainability.

+++ packages/tagger/src/providers/spotify.rs
@@ -194,7 +194,7 @@ fn convert_track(track: &FullTrack) -> Track {
         duration: Some(track.duration.num_seconds() as i32),
         file_path: None,
         track_number: Some(track.track_number as i32),
-        disc_number: Some(track.disc_number as i32),
+        disc_number: Some(track.disc_number),
         label: None,
         date: track.album.release_date.clone(),
         cover: track.album.images.first().map(|image| image.url.clone()),
